+++ src/main/java/kr/modusplant/modules/auth/social/app/error/OAuthException.java
@@ -1,4 +1,4 @@
-package kr.modusplant.modules.auth.social.error;
+package kr.modusplant.modules.auth.social.app.error;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
