+++ src/main/java/kr/modusplant/modules/auth/social/app/http/request/SocialLoginRequest.java
@@ -1,4 +1,4 @@
-package kr.modusplant.modules.auth.social.model.request;
+package kr.modusplant.modules.auth.social.app.http.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
