+++ src/main/java/kr/modusplant/modules/auth/social/app/service/KakaoAuthClient.java
@@ -1,8 +1,8 @@
-package kr.modusplant.modules.auth.social.service;
+package kr.modusplant.modules.auth.social.app.service;
 
-import kr.modusplant.modules.auth.social.dto.GoogleUserInfo;
-import kr.modusplant.modules.auth.social.error.OAuthException;
-import kr.modusplant.modules.auth.social.service.supers.SocialAuthClient;
+import kr.modusplant.modules.auth.social.app.dto.KakaoUserInfo;
+import kr.modusplant.modules.auth.social.app.error.OAuthException;
+import kr.modusplant.modules.auth.social.app.service.supers.SocialAuthClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -16,32 +16,29 @@ import org.springframework.web.client.RestClient;
 import java.util.Map;
 
 @Service
-public class GoogleAuthClient implements SocialAuthClient {
+public class KakaoAuthClient implements SocialAuthClient {
     private RestClient restClient;
-    @Value("${google.api-key}")
-    private String GOOGLE_API_KEY;
-    @Value("${google.secret}")
-    private String GOOGLE_SECRET;
-    @Value("${google.redirect-uri}")
-    private String GOOGLE_REDIRECT_URI;
+    @Value("${kakao.api-key}")
+    private String KAKAO_API_KEY;
+    @Value("${kakao.redirect-uri}")
+    private String KAKAO_REDIRECT_URI;
 
     public String getAccessToken(String code) {
         restClient = RestClient.builder()
-                .baseUrl("https://oauth2.googleapis.com")
+                .baseUrl("https://kauth.kakao.com")
                 .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_FORM_URLENCODED_VALUE)
                 .build();
 
         MultiValueMap<String,String> formData = new LinkedMultiValueMap<>();
         Map.of(
-                "code", code,
-                "client_id", GOOGLE_API_KEY,
-                "client_secret", GOOGLE_SECRET,
-                "redirect_uri", GOOGLE_REDIRECT_URI,
-                "grant_type","authorization_code"
+                "code",code,
+                "client_id",KAKAO_API_KEY,
+                "redirect_uri", KAKAO_REDIRECT_URI,
+                "grant_type", "authorization_code"
         ).forEach(formData::add);
 
         return restClient.post()
-                .uri("/token")
+                .uri("/oauth/token")
                 .body(formData)
                 .retrieve()
                 .onStatus(this::isErrorStatus, (request, response) -> {
@@ -51,19 +48,19 @@ public class GoogleAuthClient implements SocialAuthClient {
                 .get("access_token").toString();
     }
 
-    public GoogleUserInfo getUserInfo(String accessToken) {
+    public KakaoUserInfo getUserInfo(String accessToken) {
         restClient = RestClient.builder()
-                .baseUrl("https://www.googleapis.com")
+                .baseUrl("https://kapi.kakao.com")
                 .defaultHeader(HttpHeaders.AUTHORIZATION, "Bearer "+accessToken)
                 .build();
 
         return restClient.get()
-                .uri("/userinfo/v2/me")
+                .uri("/v2/user/me")
                 .retrieve()
                 .onStatus(this::isErrorStatus, (request, response) -> {
                     throw new OAuthException((HttpStatus) response.getStatusCode());
                 })
-                .body(GoogleUserInfo.class);
+                .body(KakaoUserInfo.class);
     }
 
     private boolean isErrorStatus(HttpStatusCode status) {
