+++ src/main/java/kr/modusplant/modules/auth/social/app/service/supers/SocialAuthClient.java
@@ -1,0 +1,8 @@
+package kr.modusplant.modules.auth.social.app.service.supers;
+
+import kr.modusplant.modules.auth.social.app.dto.supers.SocialUserInfo;
+
+public interface SocialAuthClient {
+    String getAccessToken(String code);
+    SocialUserInfo getUserInfo(String accessToken);
+}
