--- src/main/java/kr/modusplant/modules/auth/social/service/supers/SocialAuthClient.java
@@ -1,8 +1,0 @@
-package kr.modusplant.modules.auth.social.service.supers;
-
-import kr.modusplant.modules.auth.social.dto.supers.SocialUserInfo;
-
-public interface SocialAuthClient {
-    String getAccessToken(String code);
-    SocialUserInfo getUserInfo(String accessToken);
-}
